Extract symtable module
