+++ Makefile
@@ -3,7 +3,7 @@ CFLAGS ?= -Wall -Wextra -std=c99
 OPTFLAGS ?=
 BIN = vc
 # Core compiler sources
-CORE_SRC = src/main.c src/lexer.c src/ast.c src/parser.c src/semantic.c \
+CORE_SRC = src/main.c src/lexer.c src/ast.c src/parser.c src/symtable.c src/semantic.c \
     src/ir.c src/codegen.c src/regalloc.c src/strbuf.c src/util.c
 # Optional optimization sources
 OPT_SRC = src/opt.c
@@ -11,7 +11,7 @@ OPT_SRC = src/opt.c
 EXTRA_SRC ?=
 # Final source list
 SRC = $(CORE_SRC) $(OPT_SRC) $(EXTRA_SRC)
-HDR = include/token.h include/ast.h include/parser.h include/semantic.h \
+HDR = include/token.h include/ast.h include/parser.h include/symtable.h include/semantic.h \
     include/ir.h include/opt.h include/codegen.h include/strbuf.h \
     include/util.h
 PREFIX ?= /usr/local
